requirements.txt
transformers>=4.40.0
torch>=2.0.0
Pillow>=10.0.0
tqdm
